# rest/config/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Liverpool FC API",
    "version": "1.0.0",
    "description": "REST API for managing Liverpool Football Club data including players, matches, and more"
  },
  "servers": [
    {
      "url": "http://localhost:8989/api/v1"
    }
  ],
  "paths": {
...
    "/managers": {
      "get": {
        "summary": "Get all managers",
        "tags": [
          "Managers"
        ],
        "responses": {
          "200": {
            "description": "List of managers",
            "headers": {
              "X-Total-Count": {
                "description": "Total number of managers",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Manager"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No managers found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create new manager",
        "tags": [
          "Managers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Manager"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Manager created",
            "headers": {
              "Location": {
                "description": "URL of created manager",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ManagerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/managers/{id}": {
      "get": {
        "summary": "Get manager by ID",
        "tags": [
          "Managers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Manager ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Manager found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ManagerResponse"
                }
              }
            }
          },
          "404": {
            "description": "Manager not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update manager",
        "tags": [
          "Managers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Manager ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Manager"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Manager updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ManagerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Manager not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "patch": {
        "summary": "Partially update manager",
        "tags": [
          "Managers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Manager ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "nationality": {
                    "type": "string"
                  },
                  "dateOfBirth": {
                    "type": "string",
                    "format": "date"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "INACTIVE",
                      "SUSPENDED"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Manager updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ManagerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Manager not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "summary": "Delete manager",
        "tags": [
          "Managers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Manager ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Manager deleted"
          },
          "404": {
            "description": "Manager not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Manager": {
        "type": "object",
        "required": [
          "name",
          "nationality",
          "dateOfBirth"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "description": "Manager's full name"
          },
          "nationality": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "description": "Manager's nationality"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "description": "Manager's date of birth (must be between 18-100 years old)"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED"
            ],
            "default": "ACTIVE",
            "description": "Manager's current status"
          }
        }
      },
...
    },
    "responses": {
...
      "ManagerResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Manager"
          },
          "_links": {
            "type": "object",
            "properties": {
              "self": {
                "type": "string"
              },
              "collection": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
...
}